pyproject.toml
[tool.poetry]
name = "text-finder"
version = "0.1.0"
description = ""
authors = ["Dddarknight <9801677@gmail.com>"]
readme = "README.md"
packages = [{include = "text_finder"}]

[tool.poetry.dependencies]
python = ">=3.8,<4"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
fastapi = "^0.87.0"
pandas = "^1.5.2"
elasticsearch = {extras = ["async"], version = "^8.5.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
